packaged lexer
