+++ md_2_html/lexer/__init__.py
@@ -1,0 +1,1 @@
+from ._lexer import *
